Fix MoE test

+++ tests/kernels/test_moe.py
@@ -27,7 +27,7 @@ def torch_moe(a, w1, w2, topk_weight, topk_ids):
             out[mask] = SiluAndMul()(
                 a[mask] @ w1[i].transpose(0, 1)) @ w2[i].transpose(0, 1)
     return (out.view(B, -1, w2.shape[1]) *
-            topk_weight.view(B, -1, 1)).sum(dim=1)
+            topk_weight.view(B, -1, 1).to(out.dtype)).sum(dim=1)
 
 
 @pytest.mark.parametrize("m", [512, 222, 33, 1])
@@ -51,7 +51,7 @@ def test_fused_moe(
     score = torch.randn((m, e), device='cuda', dtype=dtype)
     triton_output = fused_moe(a, w1, w2, score, topk, renormalize=False)
 
-    score = torch.softmax(score, dim=-1)
+    score = torch.softmax(score, dim=-1, dtype=torch.float32)
     topk_weight, topk_ids = torch.topk(score, topk)
     torch_output = torch_moe(a, w1, w2, topk_weight, topk_ids)
     assert torch.allclose(triton_output, torch_output, atol=1e-2, rtol=0)
